pyproject.toml
[project]
name = "rag-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docling>=2.80.0",
    "fastapi>=0.135.1",
    "flagembedding>=1.3.5",
    "httpx>=0.28.1",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.16.1",
    "redis>=7.3.0",
    "sentence-transformers>=5.3.0",
    "uvicorn>=0.42.0",
    "vllm>=0.17.1",
]
